feat: rediseño total de UI estilo Material3 y lógica completa de calculadora funcional

+++ app/src/main/java/ni/edu/uam/pruebasunitarias/Calculadora.kt
@@ -1,11 +1,11 @@
 package ni.edu.uam.pruebasunitarias
 
 class Calculadora {
-    fun sumar(a: Int, b: Int): Int = a + b
-    fun restar(a: Int, b: Int): Int = a - b
-    fun multiplicar(a: Int, b: Int): Int = a * b
-    fun dividir(a: Int, b: Int): Int {
-        require(b != 0) { "No se puede dividir por cero" }
+    fun sumar(a: Double, b: Double): Double = a + b
+    fun restar(a: Double, b: Double): Double = a - b
+    fun multiplicar(a: Double, b: Double): Double = a * b
+    fun dividir(a: Double, b: Double): Double {
+        require(b != 0.0) { "No se puede dividir por cero" }
         return a / b
     }
 }
